requirements.txt
asgiref==3.5.1
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.13
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.21.4
factory-boy==3.2.1
Faker==13.7.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.8
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-environ==0.4.54
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
zipp==3.8.0
